refactor: move sample export logic to eth_defi.vault.sample_export module
Replace importlib hack with a proper module import. The script is now a
thin CLI wrapper.
